Updated Bug Fixing
STMap: The STMap code isn't capturing the title name correctly.
Refresh: When refreshing, the snippet type and list reset, which causes the expected effect but creates the inconvenience of having to reselect the desired code.
Append Next Line Behavior: When appending, the code is added to the wrangler jumping to the next line. This is the expected behavior, but when there isn't text, it jumps to the next line anyway.
Append Code Behavior: When appending, the code is added to the first selected wrangler instead of the current selected.